+++ app/src/organisms/ODD/QuickTransferFlow/utils/createQuickTransferFile.ts
@@ -4,8 +4,12 @@ import {
   transfer,
   distribute,
   getWasteChuteAddressableAreaNamePip,
+  pythonImports,
+  pythonMetadata,
+  pythonRequirements,
 } from '@opentrons/step-generation'
 import { generateQuickTransferArgs } from './'
+import { pythonCommands } from './pythonCommands'
 import {
   FLEX_ROBOT_TYPE,
   FLEX_STANDARD_DECKID,
@@ -254,4 +258,38 @@ export function createQuickTransferFile(
     [protocolContents],
     `${protocolBase.metadata.protocolName}.json`
   )
+}
+
+export function createQuickTransferPythonFile(
+  quickTransferState: QuickTransferSummaryState,
+  deckConfig: DeckConfiguration,
+  protocolName?: string
+): string {
+  const sourceLabwareName = quickTransferState.source.metadata.displayName
+  let destinationLabwareName = sourceLabwareName
+  if (quickTransferState.destination !== 'source') {
+    destinationLabwareName = quickTransferState.destination.metadata.displayName
+  }
+  const fileMetadata = {
+    protocolName:
+      protocolName ?? `Quick Transfer ${quickTransferState.volume}µL`,
+    description: `This quick transfer moves liquids from a ${sourceLabwareName} to a ${destinationLabwareName}`,
+    source: 'Quick Transfer',
+    version: '1.1.0',
+    category: null,
+    subcategory: null,
+    tags: [],
+  }
+
+  const protocolContents =
+    [
+      pythonImports(),
+      pythonMetadata(fileMetadata),
+      pythonRequirements(FLEX_ROBOT_TYPE),
+      pythonCommands(quickTransferState, deckConfig),
+    ]
+      .filter(section => section)
+      .join('\n\n') + '\n'
+
+  return protocolContents
 }
