+++ app/src/organisms/ODD/QuickTransferFlow/utils/generateQuickTransferArgs.ts
@@ -32,7 +32,11 @@ import type {
   AdditionalEquipmentEntities,
 } from '@opentrons/step-generation'
 
-type MoveLiquidStepArgs = ConsolidateArgs | DistributeArgs | TransferArgs | null
+export type MoveLiquidStepArgs =
+  | ConsolidateArgs
+  | DistributeArgs
+  | TransferArgs
+  | null
 
 const uuid: () => string = uuidv1
 const adapter96ChannelDefUri = 'opentrons/opentrons_flex_96_tiprack_adapter/1'
