+++ protocol-designer/src/file-data/selectors/utils.ts
@@ -11,7 +11,6 @@ import type {
   LoadModuleCreateCommand,
   LoadPipetteCreateCommand,
   PipetteName,
-  PipetteV2Specs,
 } from '@opentrons/shared-data'
 import type {
   LabwareEntities,
@@ -237,14 +236,4 @@ export const getLabwareLoadInfo = (
     }),
     {}
   )
-}
-
-const DEFAULT_LIQUID_TYPE = 'default'
-//  Flex pipette api names are different from pipetteName
-//  p1000_multi_flex -> flex_8channel_1000
-//  we do not need to worry about -_em pipette in PD
-export const getFlexNameConversion = (pipetteSpec: PipetteV2Specs): string => {
-  const channels = pipetteSpec.channels
-  const maxVolume = pipetteSpec.liquids[DEFAULT_LIQUID_TYPE].maxVolume
-  return `flex_${channels}channel_${maxVolume}`
 }
