+++ step-generation/src/__tests__/generatePythonFileUtils.test.ts
@@ -23,7 +23,7 @@ import {
   getLoadWasteChute,
   pythonMetadata,
   pythonRequirements,
-} from '../selectors/pythonFile'
+} from '../utils'
 import type { LabwareDefinition2 } from '@opentrons/shared-data'
 import type {
   AdditionalEquipmentEntities,
